+ fix BitStore write Integer.MIN_VALUE

+++ src/test/java/de/longri/serializable/BitStoreTest.java
@@ -102,21 +102,20 @@ public class BitStoreTest {
         obj.value1 = -36;
         obj.value2 = 117;
         obj.value3 = 36;
-        obj.value4 = Short.MIN_VALUE;
-        obj.value5 = Short.MAX_VALUE;
+        obj.value4 = Integer.MIN_VALUE;
+        obj.value5 = Integer.MAX_VALUE;
         obj.value6 = 0;
-        obj.value7 = Short.MIN_VALUE + 1;
-        obj.value8 = Short.MAX_VALUE - 1;
+        obj.value7 = Integer.MIN_VALUE + 1;
+        obj.value8 = Integer.MAX_VALUE - 1;
 
-        obj.value9 = Short.MIN_VALUE + (Short.MAX_VALUE / 2);
-        obj.value10 = Short.MAX_VALUE - (Short.MAX_VALUE / 2);
+        obj.value9 = Integer.MIN_VALUE + (Integer.MAX_VALUE / 2);
+        obj.value10 = Integer.MAX_VALUE - (Integer.MAX_VALUE / 2);
 
-        obj.value11 = Short.MIN_VALUE + (Short.MAX_VALUE / 4);
-        obj.value12 = Short.MAX_VALUE - (Short.MAX_VALUE / 4);
+        obj.value11 = Integer.MIN_VALUE + (Integer.MAX_VALUE / 4);
+        obj.value12 = Integer.MAX_VALUE - (Integer.MAX_VALUE / 4);
 
-        obj.value13 = Short.MIN_VALUE + (Short.MAX_VALUE / 8);
-        obj.value14 = Short.MAX_VALUE - (Short.MAX_VALUE / 8);
-
+        obj.value13 = Integer.MIN_VALUE + (Integer.MAX_VALUE / 8);
+        obj.value14 = Integer.MAX_VALUE - (Integer.MAX_VALUE / 8);
 
         TestObjectInteger obj2 = new TestObjectInteger();
 
@@ -136,20 +135,20 @@ public class BitStoreTest {
         obj.value1 = -36;
         obj.value2 = 117;
         obj.value3 = 36;
-        obj.value4 = Short.MIN_VALUE;
-        obj.value5 = Short.MAX_VALUE;
+        obj.value4 = Long.MIN_VALUE;
+        obj.value5 = Long.MAX_VALUE;
         obj.value6 = 0;
-        obj.value7 = Short.MIN_VALUE + 1;
-        obj.value8 = Short.MAX_VALUE - 1;
+        obj.value7 = Long.MIN_VALUE + 1;
+        obj.value8 = Long.MAX_VALUE - 1;
 
-        obj.value9 = Short.MIN_VALUE + (Short.MAX_VALUE / 2);
-        obj.value10 = Short.MAX_VALUE - (Short.MAX_VALUE / 2);
+        obj.value9 = Long.MIN_VALUE + (Long.MAX_VALUE / 2);
+        obj.value10 = Long.MAX_VALUE - (Long.MAX_VALUE / 2);
 
-        obj.value11 = Short.MIN_VALUE + (Short.MAX_VALUE / 4);
-        obj.value12 = Short.MAX_VALUE - (Short.MAX_VALUE / 4);
+        obj.value11 = Long.MIN_VALUE + (Long.MAX_VALUE / 4);
+        obj.value12 = Long.MAX_VALUE - (Long.MAX_VALUE / 4);
 
-        obj.value13 = Short.MIN_VALUE + (Short.MAX_VALUE / 8);
-        obj.value14 = Short.MAX_VALUE - (Short.MAX_VALUE / 8);
+        obj.value13 = Long.MIN_VALUE + (Long.MAX_VALUE / 8);
+        obj.value14 = Long.MAX_VALUE - (Long.MAX_VALUE / 8);
 
 
         TestObjectLong obj2 = new TestObjectLong();
